requirements.txt
setuptools==19.4
docker-py==1.7.2
python-dockercloud==1.0.2
requests==2.7.0
six==1.9.0
websocket-client==0.32.0
